Provide fixes for missing type exposes from same module
If the type is declared in another module then we can't expose it here,
the author may choose to create and expose a type alias but that's not
something the rule should do automatically as there may already be a
type alias exposed somewhere else that could be used.

This patch also fixes an issue where types exposed on import did not
report with their fully qualified name.